refactor(data-pipeline): rename surge_*.py model/data modules to vst_*
Follow-up to #1602, which renamed the Surge* dataset/datamodule/model classes
to VST* but left the module filenames as surge_*. Renames the five module files
to vst_* so paths match their class names, and migrates the path-keyed lint
exclusion entries (ruff per-file-ignores, pyright excludes, pre-commit pyright/
interrogate excludes, pydoclint baseline) to the new paths — the grandfathered
errors stay suppressed under vst_* rather than being un-exempted by the rename.

Thin re-export shims remain at each old surge_* path so archived W&B run-config
_target_s (e.g. surge_ff_module.SurgeFeedForwardModule) keep resolving;
test_vst_module_aliases.py now pins both the shim resolution and the Surge*->VST*
alias identity across the rename.

Closes #1664
