+++ pyproject.toml
@@ -394,10 +394,10 @@ known-third-party = ["wandb"]
 "src/synth_setter/models/components/vae.py" = ["E402", "T201", "D102", "D103", "D107"]
 "src/synth_setter/models/components/vector_field.py" = ["D102", "D103", "D107"]
 "src/synth_setter/models/ksin_ff_module.py" = ["F821", "D102", "D103", "D107"]
-"src/synth_setter/models/surge_ff_module.py" = ["D102", "D103", "D107"]
+"src/synth_setter/models/vst_ff_module.py" = ["D102", "D103", "D107"]
 "src/synth_setter/models/ksin_flow_matching_module.py" = ["D102", "D103", "D107"]
-"src/synth_setter/models/surge_flow_matching_module.py" = ["D102", "D103", "D107"]
-"src/synth_setter/models/surge_flowvae_module.py" = ["D102", "D103", "D107"]
+"src/synth_setter/models/vst_flow_matching_module.py" = ["D102", "D103", "D107"]
+"src/synth_setter/models/vst_flowvae_module.py" = ["D102", "D103", "D107"]
 "src/synth_setter/cli/train.py" = ["E402", "D102", "D103", "D107"]
 "src/synth_setter/utils/__init__.py" = ["F401"]
 "src/synth_setter/utils/callbacks.py" = ["F821", "T201", "ANN001", "D102", "D103", "D107"]
