+++ pyrightconfig.json
@@ -15,7 +15,7 @@
     "src/synth_setter/data/fm_datamodule.py",
     "src/synth_setter/data/kosc_datamodule.py",
     "src/synth_setter/data/ot.py",
-    "src/synth_setter/data/surge_datamodule.py",
+    "src/synth_setter/data/vst_datamodule.py",
     "src/synth_setter/data/vst/core.py",
     "src/synth_setter/data/vst/generate_vst_dataset.py",
     "src/synth_setter/data/vst/param_spec.py",
@@ -27,9 +27,9 @@
     "src/synth_setter/models/components/vector_field.py",
     "src/synth_setter/models/ksin_ff_module.py",
     "src/synth_setter/models/ksin_flow_matching_module.py",
-    "src/synth_setter/models/surge_ff_module.py",
-    "src/synth_setter/models/surge_flow_matching_module.py",
-    "src/synth_setter/models/surge_flowvae_module.py",
+    "src/synth_setter/models/vst_ff_module.py",
+    "src/synth_setter/models/vst_flow_matching_module.py",
+    "src/synth_setter/models/vst_flowvae_module.py",
     "src/synth_setter/cli/train.py",
     "src/synth_setter/utils/callbacks.py",
     "src/synth_setter/utils/logging_utils.py",
