+++ src/synth_setter/configs/datamodule/vst.yaml
@@ -1,4 +1,4 @@
-_target_: synth_setter.data.surge_datamodule.VSTDataModule
+_target_: synth_setter.data.vst_datamodule.VSTDataModule
 dataset_root: ${paths.output_dir}/data
 download_dataset_root_uri: null
 use_saved_mean_and_variance: true
